update algorithm bug in NCLBoard about total continuous number when not contain space block Create BigPosInt for big positive integer when puzzle size larger then 5 Create UnitTest for BigPosInt

+++ 15_Puzzle/NCLBoard.h
@@ -40,8 +40,8 @@ public:
 	NCLBoard(const NCLBoard& rhs) : Board(rhs) {}
 	virtual ~NCLBoard() {};
 	bool IsTurnEnd(); // check if this board is a turn end 
-	unsigned long GetTotalContinuousNumber(const bool& containSpace, const int& partial) const; // return continuous number (all are the same)
-	unsigned long GetTotalContinuousNumber(const bool& containSpace) const; // return continuous number (all are the same)
+	unsigned long long GetTotalContinuousNumber(const bool& containSpace, const int& partial) const; // return continuous number (all are the same)
+	unsigned long long GetTotalContinuousNumber(const bool& containSpace) const; // return continuous number (all are the same)
 	ContinuousNumber CheckContinuous(const bool& containSPACE) const; // check full continuous in all directions
 	ContinuousNumber CheckContinuous(const bool& containSPACE, const int& length) const; // check partial continuous in all directions
 	string ToString();
